Changes after reveiw
