+++ GovUkDesignSystemComponents/Enums/CellFormat.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace GovUkDesignSystem.GovUkDesignSystemComponents.Enums
+{
+    public enum Format
+    { 
+        Numeric
+    }
+}
